+++ versions/3.0.1/views/Pages.php
@@ -1,9 +1,14 @@
 <div class="top-flex d-flex justify-content-between">
     <h1><?php $cms->PrintTranslate('Pages'); ?></h1>
     <div class="buttons">
-        <button class="cms-btn btn-iconed" @click="toggleExpand()"> <?php $cms->PrintTranslate('Expand'); ?> <em class="fas fa-eye"></em></button>    
-        <button class="cms-btn btn-iconed" @click="deleteAllPages()"> <?php $cms->PrintTranslate('DeleteAllPages'); ?> <em class="fas fa-trash"></em></button>
-        <button class="cms-btn btn-iconed" @click="createPage()"> <?php $cms->PrintTranslate('CreatePage'); ?> <em class="fas fa-plus"></em></button>
+        <button class="cms-btn btn-iconed" @click="toggleExpand()"> <?php $cms->PrintTranslate('Expand'); ?> <em class="fas fa-eye"></em></button> 
+        <?php if($cms->HasAtLeast("Admin")): ?>
+            <button class="cms-btn btn-iconed" @click="deleteAllPages()"> <?php $cms->PrintTranslate('DeleteAllPages'); ?> <em class="fas fa-trash"></em></button>
+        <?php endif; ?>
+
+        <?php if($cms->HasAtLeast("Moderator")): ?>
+            <button class="cms-btn btn-iconed" @click="createPage()"> <?php $cms->PrintTranslate('CreatePage'); ?> <em class="fas fa-plus"></em></button>
+        <?php endif; ?>
     </div>
 </div>
 
@@ -24,9 +29,11 @@
             <?php $cms->IncludeComponent("PageItem", [ "page" => $page, "depth" => 1 ]); ?>
         <?php endforeach; ?>
 
-        <div class="d-flex justify-content-center">
-            <button class="cms-btn mt-2" @click="createPage()">  <em class="fas fa-plus no"></em></button>
-        </div>
+        <?php if($cms->HasAtLeast("Moderator")): ?>
+            <div class="d-flex justify-content-center">
+                <button class="cms-btn mt-2" @click="createPage()">  <em class="fas fa-plus no"></em></button>
+            </div>
+        <?php endif; ?>
         
     <?php endif; ?>
 </div>
